Split monolithic logic into dedicated modules
